backend/requirements.txt
Django
# django-crispy-forms
pyresparser
gunicorn
spacy==2.3.5
djangorestframework